requirements.txt
# Stock Analysis
yfinance>=0.2.36
yahoo-fin>=0.3.0

# AI Analysis
google-generativeai>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Scheduling & Email
schedule>=1.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
markdown>=3.5.0
